test(eval): split the probe tests under the 500-line taste ceiling
The single test module reached 531 lines, which the taste count ratchet
reports as a new error-severity violation (567 > baseline 566). Split by
subject rather than by mechanics: plan construction moves to its own module,
the recorded output shapes and the argv-dispatching fake runner move to a
shared fixtures module, and execution keeps the original file.

Same 36 cases, same assertions. The 21 guard-removal mutations were rerun
against the split layout with __pycache__ cleared between each mutation and
its rerun: all 21 killed, the inverted control survived, and the restore was
byte-compared.

Refs #5423
